Update display constraints and logic
Created a function to check calculator display and adjust
text size as needed. Also set a hard character limit of 47.

Updated logic to reset when Infinity is reached as well as
when the result exceeds 47 characters. Additionally, when
entering a factorial of 153 or higher the calculation stops
as it would result in infinity anyway.

Updated logic to round results with decimals to the nearest
10^20.